pyproject.toml
[project]
name = "therpay-finetune"
version = "0.1.0"
description = "Privacy-first therapeutic coaching model fine-tuning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx==0.28.1",
    # Pinned for GPT-5 responses API compatibility
    "openai==2.14.0",
    "pydantic==2.12.5",
    "python-dotenv>=1.0.0",
    "pyyaml==6.0.3",
    "tenacity==9.1.2",
]

[dependency-groups]
dev = [
    "pre-commit==4.5.1",
    "ruff==0.14.10",
    "ty==0.0.6",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
]
